remove BuilderTestBase class
